requirements.txt
asgiref==3.3.2
Django==3.2.8
djangorestframework==3.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.9